requirements.txt
Django==4.2.5
djangorestframework==3.14.0
drf-yasg==1.21.7
drf-nested-routers==0.93.4
# celery[redis]==5.3.4
# flower==2.0.1
djoser==2.2.0
psycopg2==2.8.6